:lipstick: create style files for RecommendationScreen EmotionList
