+++ src/screens/RecommendationScreen.tsx
@@ -7,6 +7,7 @@ import {
 	Dimensions
 } from "react-native";
 
+import styles from "@/src/styles/RecommendationScreenStyle";
 import Emotion from "@/src/models/Emotion";
 import MovieList from "../components/MovieList";
 import EmotionsList from "../components/EmotionsList";
@@ -216,20 +217,3 @@ export default function RecommendationScreen() {
 		</SafeAreaView>
 	);
 }
-
-const styles = StyleSheet.create({
-	container: {
-		flex: 1,
-		backgroundColor: "#0A1E38"
-	},
-	overlayContainer: {
-		position: "absolute",
-		top: 0,
-		left: 0,
-		right: 0,
-		bottom: 0,
-		width: "100%",
-		height: "100%",
-		zIndex: 1
-	}
-});
