+++ src/styles/EmotionListStyle.ts
@@ -1,0 +1,21 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+	container: {
+		flex: 1,
+		width: "100%",
+		height: "100%",
+		paddingHorizontal: 16
+	},
+	scrollContainer: {
+		paddingVertical: 20,
+		paddingBottom: 40
+	},
+	grid: {
+		flexDirection: "row",
+		flexWrap: "wrap",
+		justifyContent: "space-between"
+	}
+});
+
+export default styles;
